[SPIKE] Reorganize contentful action to take an array of actions to run
