+++ action.yml
@@ -1,6 +1,8 @@
 name: 'Contentful Action'
 description: 'Run a Migration against your Contentful space'
 inputs:
+  actions:
+    description: "The actions to run on this workflow. Comma separated string. Example: `createEnvironment, createCDAToken, applyMigration, updateAlias, cleanUpEnvironments`"
   delete_feature:
     description: "The head branch will be deleted after it was merged"
   set_alias:
