+++ src/action/createCDAToken.ts
@@ -1,0 +1,32 @@
+import type { Environment, Space } from "contentful-management";
+import { Logger } from "../utils";
+
+export default async function ({ space, tokenKeyName, environment }: { space: Space, tokenKeyName: string, environment: Environment }) {
+    const spaceKeys = await space.getApiKeys();
+
+    const exists = spaceKeys.items.some(item => item.name === tokenKeyName)
+
+    if (exists) {
+        Logger.log(`CDA token ${tokenKeyName} is already created`);
+    } else {
+        Logger.log(`Creating new CDA token "${tokenKeyName}" for ephemeral environment "${environment.sys.id}"...`);
+
+        try {
+            const key = await space.createApiKey({
+                name: tokenKeyName,
+                environments: [{
+                    sys: {
+                        type: "Link",
+                        linkType: "Environment",
+                        id: environment.sys.id,
+                    },
+                }],
+            })
+
+            Logger.success("CDA token has been created");
+        } catch (err) {
+            Logger.warn("unable to create ephemeral token");
+            Logger.verbose(err)
+        }
+    }
+}
